# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Hyperlofy Hyperlocal Delivery Platform API
  description: Production-Grade Modular Monolith Backend for a real-time hyperlocal delivery network.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Create a customer, agent, or admin profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      summary: Authenticate and log in
      description: Returns JWT access token and refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /orders:
    post:
      summary: Create a new delivery order
      description: Restricts to Authenticated Customers.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /orders/{id}:
    get:
      summary: Fetch order details
      description: Order ownership verified (Customer, Agent or Admin).
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /wallets/me:
    get:
      summary: Retrieve authenticated user wallet details
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
  /payments/pay/{orderId}:
    post:
      summary: Complete delivery fee wallet payment
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Payment Authorized & Captured

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - phoneNumber
        - role
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        role:
          type: string
          enum: [CUSTOMER, AGENT, ADMIN, SUPER_ADMIN]
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        email:
          type: string
        role:
          type: string
    OrderRequest:
      type: object
      required:
        - customerId
        - zoneId
        - storeName
        - storeLatitude
        - storeLongitude
        - deliveryAddress
        - deliveryLatitude
        - deliveryLongitude
        - itemsDesc
      properties:
        customerId:
          type: string
          format: uuid
        zoneId:
          type: string
          format: uuid
        storeName:
          type: string
        storeLatitude:
          type: number
        storeLongitude:
          type: number
        deliveryAddress:
          type: string
        deliveryLatitude:
          type: number
        deliveryLongitude:
          type: number
        itemsDesc:
          type: string
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        customerName:
          type: string
        agentId:
          type: string
          format: uuid
        agentName:
          type: string
        zoneId:
          type: string
          format: uuid
        zoneName:
          type: string
        storeName:
          type: string
        deliveryAddress:
          type: string
        distanceKm:
          type: number
        deliveryFee:
          type: number
        orderStatus:
          type: string
        otpCode:
          type: string
        createdAt:
          type: string
          format: date-time
    WalletResponse:
      type: object
      properties:
        walletId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        balance:
          type: number
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
        transactionType:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time